+++ source/extensions/isaacsim.ros2.nodes/python/tests/test_service_prim.py
@@ -19,7 +19,7 @@ import numpy as np
 import omni.graph.core as og
 import omni.kit.test
 import omni.usd
-from isaacsim.core.utils.stage import create_new_stage_async
+from isaacsim.core.experimental.utils import stage as stage_utils
 from isaacsim.ros2.core.impl.ros2_test_case import ROS2TestCase
 from pxr import Sdf
 
@@ -27,7 +27,7 @@ from pxr import Sdf
 class TestRos2ServicePrim(ROS2TestCase):
     async def setUp(self):
         await super().setUp()
-        await create_new_stage_async()
+        await stage_utils.create_new_stage_async()
 
     async def tearDown(self):
         await super().tearDown()